# src/main/resources/api-spec.yaml
openapi: 3.0.2
info:
    title: The CRM Service API
    version: 1.0.0
    description: A brand new API with no content.  Go nuts!
paths:
    /v1/customer:
        get:
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CustomerList'
                    description: OK
            security:
                -
                    TheCRMServiceSecurityScheme:
                        - user
                        - admin
            operationId: getCustomers
            summary: Get all customer
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Customer'
                required: true
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Customer'
                    description: OK
            operationId: createCustomer
            summary: Create a customer
    '/v1/customer/{id}':
        get:
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Customer'
                    description: OK
            security:
                -
                    TheCRMServiceSecurityScheme:
                        - user
                        - admin
            operationId: getCustomer
            summary: Get customer
        put:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Customer'
                required: true
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Customer'
                    description: OK
            operationId: updateCustomer
            summary: Update customer
        delete:
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Customer'
                    description: OK
            operationId: deleteCustomer
            summary: Delete customer
        parameters:
            -
                name: id
                description: Customer ID
                schema:
                    type: string
                in: path
                required: true
    /v1/user:
        get:
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserList'
                    description: OK
            security:
                -
                    TheCRMServiceSecurityScheme:
                        - admin
            operationId: getUsers
            summary: Get users
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/User'
                required: true
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                    description: OK
            operationId: createUser
            summary: Create user
    '/v1/user/{id}':
        get:
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                    description: OK
            security:
                -
                    TheCRMServiceSecurityScheme:
                        - admin
            operationId: getUser
            summary: Get user
        put:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/User'
                required: true
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                    description: OK
            operationId: updateUser
            summary: Update user
        delete:
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                    description: OK
            operationId: deleteUser
            summary: Delete user
        parameters:
            -
                name: id
                description: User ID
                schema:
                    type: string
                in: path
                required: true
    /v1/login/oauth2/github/authorize:
        get:
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GenericResponse'
                    description: OK
            operationId: githubAuthorize
            summary: Authorize an user using Github OAuth2
        parameters:
            -
                name: code
                description: ''
                schema:
                    type: string
                in: query
                required: false
            -
                name: state
                description: ''
                schema:
                    type: string
                in: query
                required: false
            -
                name: response_type
                description: ''
                schema:
                    type: string
                in: query
            -
                name: client_id
                description: ''
                schema:
                    type: string
                in: query
                required: false
            -
                name: scope
                description: ''
                schema:
                    type: string
                in: query
            -
                name: redirect_uri
                description: ''
                schema:
                    type: string
                in: query
                required: false
    /v1/login/oauth2/github:
        get:
            responses:
                '200':
                    content:
                        text/xml:
                            schema:
                                type: string
                    description: OK
                '302':
                    content:
                        text/xml:
                            schema:
                                type: string
                    description: Redirect page
            operationId: githubLogin
            summary: Redirect user to Github's authentication server
components:
    schemas:
        CustomerList:
            title: Root Type for CustomerListDTO
            description: ''
            type: array
            items:
                $ref: '#/components/schemas/Customer'
            example:
                -
                    id: dummyId
                    name: dummyName
                    surname: dummySurname
                    lastUpdateBy: fakeUser
                    createdBy: fakeUser
        Customer:
            title: Root Type for CustomerDTO
            description: ''
            required:
                - name
                - surname
                - id
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                surname:
                    type: string
                lastUpdateBy:
                    type: string
                createdBy:
                    type: string
                imageUrl:
                    description: ''
                    type: string
            example:
                id: dummyId
                name: dummyName
                surname: dummySurname
                lastUpdateBy: fakeDate
                createdBy: fakeUser
        User:
            title: Root Type for User
            description: ''
            required:
                - id
                - role
                - name
            type: object
            properties:
                id:
                    type: string
                role:
                    description: ''
                    enum:
                        - ADMIN
                        - USER
                    type: string
                name:
                    description: ''
                    type: string
            example:
                id: fakeId
                role: ADMIN
                name: Fake Name
        UserList:
            title: Root Type for UserList
            description: ''
            type: array
            items:
                $ref: '#/components/schemas/User'
            example:
                -
                    id: fakeId
                    isAdmin: false
        GenericResponse:
            title: Root Type for GenericResponse
            description: ''
            required:
                - message
            type: object
            properties:
                message:
                    type: string
                errors:
                    description: ''
                    type: array
                    items:
                        $ref: '#/components/schemas/Error'
            example:
                message: Fake message
        Error:
            title: Root Type for Error
            description: ''
            required:
                - error
                - field
            type: object
            properties:
                field:
                    type: string
                error:
                    type: string
            example:
                field: FakeField
                error: FakeError
    securitySchemes:
        TheCRMServiceSecurityScheme:
            flows:
                authorizationCode:
                    authorizationUrl: 'https://d00b-185-237-136-194.eu.ngrok.io/api/login/oauth2/authorize/github'
                    tokenUrl: 'https://d00b-185-237-136-194.eu.ngrok.io/api/oauth2/token/github'
                    refreshUrl: 'https://d00b-185-237-136-194.eu.ngrok.io/api/oauth2/refresh/github'
                    scopes:
                        user: ''
                        admin: ''
            type: oauth2
security:
    -
        TheCRMServiceSecurityScheme:
            - user
            - admin
tags:
    -
        name: CRMApi
        description: ''